agent/requirements.txt
uvicorn
pydantic-ai-slim[ag-ui]
python-dotenv
